Changing goals memorization from int to double so we can memorize average rewards.

+++ src/main/java/csironi/ggp/course/MCTS/backpropagation/OldBackpropagationStrategy.java
@@ -12,6 +12,6 @@ import java.util.List;
  */
 public interface OldBackpropagationStrategy {
 
-	public void update(List<Integer> goals);
+	public void update(List<Double> goals);
 
 }
